Switch to ASL for Future Usage
All waypoints and obstacles are no w based on ASL altitude - except for air drop. This allows for us to directly use the interop code provided by competition - which greatly simplifies the interop integration process.

Additional Improvement: Fly Zone is automatically updated based on imported information. These were hard-coded in the past since we know the details in the rules.